Load population and death data, disaggregated by year, state, sex and age.

get_death_population function that allows the user to get a demographic dataset of the population and deaths based on the given parameters.

Usage

get_death_population(
  v_state = "National",
  v_year = "2021",
  v_sex = "Total",
  v_age = c(0, 5, 15, 25, 45, 55, 65, 70),
  age_groups = TRUE
)

Arguments

v_state

Character vector specifying the state(s) that the function will return.

v_year

Numeric vector that specifies the year(s) to return. Must have numbers between 1970 and 2050. Default is 2021.

v_sex

Character vector that selects sex. Options: Female, Male and Total.

v_age

Numeric vector that specifies the age(s) to return.

age_groups

Specifies whether to aggregate the output by age groups.

Value

A demographic dataset based on the selected parameters.

Examples

get_death_population(v_state = c("Aguascalientes", "Campeche"),
                     v_year = c(2021, 2046),
                     v_sex = c("Female", "Total"),
                     age_groups = TRUE)
#> # A tibble: 64 x 7
#>     year state          CVE_GEO sex    age_group population deaths
#>    <int> <chr>            <dbl> <chr>  <fct>          <dbl>  <dbl>
#>  1  2021 Aguascalientes       1 Female [0,5]          73217    124
#>  2  2021 Aguascalientes       1 Female (5,15]        129266     27
#>  3  2021 Aguascalientes       1 Female (15,25]       129438     61
#>  4  2021 Aguascalientes       1 Female (25,45]       216476    225
#>  5  2021 Aguascalientes       1 Female (45,55]        82038    299
#>  6  2021 Aguascalientes       1 Female (55,65]        55209    508
#>  7  2021 Aguascalientes       1 Female (65,70]        17276    319
#>  8  2021 Aguascalientes       1 Female (70,Inf]       29851   1733
#>  9  2021 Aguascalientes       1 Total  [0,5]         149230    281
#> 10  2021 Aguascalientes       1 Total  (5,15]        264282     76
#> # ... with 54 more rows

get_death_population(v_state = "National",
                     v_year = c(2021, 2046, 2050),
                     v_sex = "Total",
                     age_groups= FALSE)
#> # A tibble: 24 x 7
#>     year state    CVE_GEO sex     age population deaths
#>    <int> <chr>      <dbl> <chr> <int>      <dbl>  <dbl>
#>  1  2021 National       0 Total     0    2127321  26647
#>  2  2021 National       0 Total     5    2193584    543
#>  3  2021 National       0 Total    15    2228965   1209
#>  4  2021 National       0 Total    25    2153706   3066
#>  5  2021 National       0 Total    45    1630861   5622
#>  6  2021 National       0 Total    55    1265172   9809
#>  7  2021 National       0 Total    65     809988  14579
#>  8  2021 National       0 Total    70     579843  15892
#>  9  2046 National       0 Total     0    1743438   9001
#> 10  2046 National       0 Total     5    1813726    130
#> # ... with 14 more rows